feat(admin): 398 - filtres tree - passage de la version optimisée sur autre branche
