+++ admin/src/components/filters-system-v2/tree-filter/TreeFilter.tsx
@@ -1,6 +1,10 @@
-import cx from "classnames";
-import React from "react";
-import { useTreeFilter } from "./TreeFilterContext";
+import ButtonPrimary from "@/components/ui/buttons/ButtonPrimary";
+import React, { useRef, useState } from "react";
+import { InnerLevelTreeFilter } from "./InnerLevelTreeFilter";
+import { SelectedValues } from "./SelectedValues";
+import { TreeFilterProvider } from "./TreeFilterContext";
+import { TreeNodeFilterType } from "./TreeNodeFilter";
+import { useClickOutside } from "./hooks/useClickOutside";
 
 export interface TreeFilterProps {
   id: string;
@@ -9,64 +13,29 @@ export interface TreeFilterProps {
   showSelectedValuesCount?: boolean;
 }
 
-export interface TreeNodeFilterType {
-  checked?: boolean;
-  label: string;
-  value: string;
-  childIds?: string[];
-  parentId?: string;
-  isLeaf: boolean;
-  isRoot: boolean;
-  groupKey: string;
-}
+export function TreeFilter({ id, treeFilter, showSelectedValues, showSelectedValuesCount }: TreeFilterProps) {
+  const [isOpen, setIsOpen] = useState(false);
+  const ref = useRef<HTMLDivElement>(null);
+  useClickOutside(ref, () => setIsOpen(false));
 
-export interface TreeNodeFilterProps {
-  item: TreeNodeFilterType;
-  className?: string;
-}
-export function TreeNodeFilter({ nodeId }: { nodeId: string }) {
-  const { onCheckboxClick, getItemState, isIndeterminate, getSelectedChildrenCount, id, getNode, showSelectedValuesCount } = useTreeFilter();
-  const item = getNode(nodeId);
-  if (!item) return null;
-
-  const selectedChildrenCount = getSelectedChildrenCount(nodeId);
   return (
-    <div
-      className="w-full h-full"
-      onClick={(e) => {
-        if (item.isLeaf) {
-          onCheckboxClick(nodeId);
-        }
-      }}>
-      <div className={cx("flex items-center gap-2 pl-2 full-width")}>
-        {!item.isRoot && (
-          <input
-            className="hover:cursor-pointer"
-            key={`input-${id}-${item.value}`}
-            type="checkbox"
-            checked={getItemState(nodeId)}
-            ref={(input) => {
-              if (input) {
-                input.indeterminate = isIndeterminate(nodeId);
-              }
-            }}
-            onChange={() => {
-              if (!item.isLeaf) {
-                onCheckboxClick(nodeId);
-              }
-            }}
-          />
-        )}
-
-        <div className="flex items-center gap-2">
-          <span>{item.label}</span>
-          {showSelectedValuesCount && !!selectedChildrenCount && (
-            <div className="flex h-6 w-6 items-center justify-center rounded-full bg-indigo-100 text-xs font-normal text-blue-600">{selectedChildrenCount}</div>
+    <TreeFilterProvider flatTree={treeFilter} id={id} showSelectedValuesCount={showSelectedValuesCount}>
+      <div className="flex flex-col" ref={ref}>
+        <div className="relative flex-row">
+          <div className="flex items-center gap-2 p-2">
+            <ButtonPrimary onClick={() => setIsOpen(!isOpen)} className="h-[50px] w-[150px]">
+              Filtres
+            </ButtonPrimary>
+          </div>
+          {isOpen && treeFilter["0"] && (
+            <div className="absolute left-0 z-10 mt-2 w-64 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5">
+              <div className="px-3 pt-3 text-xs font-light leading-5 text-gray-500">{treeFilter["0"].label}</div>
+              <div className="flex flex-col gap-1 py-2">{treeFilter["0"].childIds?.map((childId) => <InnerLevelTreeFilter key={`root-${id}-${childId}`} nodeId={childId} />)}</div>
+            </div>
           )}
         </div>
-
-        {item.childIds && item.childIds.length > 0 && <span className="absolute right-4">{">"}</span>}
+        {showSelectedValues && <SelectedValues />}
       </div>
-    </div>
+    </TreeFilterProvider>
   );
 }
