--- admin/src/components/filters-system-v2/tree-filter/index.optimized.tsx
@@ -1,4 +1,0 @@
-export { TreeFilterWithoutHeadlessUi } from "./TreeFilterWithoutHeadlessUi.optimized";
-export { TreeFilterProvider, useTreeFilter } from "./TreeFilterContext.optimized";
-export type { TreeFilterProps, TreeNodeFilterType } from "./TreeFilter.optimized";
-export { TreeNodeFilter } from "./TreeFilter.optimized";
