+++ admin/src/scenes/develop/DesignSystemPage.jsx
@@ -13,7 +13,6 @@ import ModalExamples from "./components/Modal";
 import ProfilePicExamples from "./components/ProfilePic";
 import Colors from "./components/Colors";
 import TreeFilterExample from "./TreeFilterExample";
-import { TreeFilterPerformanceTest } from "@/components/filters-system-v2/tree-filter/PerformanceTest";
 
 export default function DesignSystemPage() {
   const [values, setValues] = React.useState({
@@ -218,11 +217,8 @@ export default function DesignSystemPage() {
         ]}
         button={[<Button key={"button-1"} title={"Primary base"} />, <Button key={"button-2"} title={"Primary base"} loading />]}
       />
-      <Container title="Filtres de type Tree WIP">
+      <Container title="Filtres de type Tree">
         <TreeFilterExample />
-      </Container>
-      <Container title="Filtres de type Tree Performance Test">
-        <TreeFilterPerformanceTest />
       </Container>
       <Container title="Champs simples (InputText)">
         <div className="grid grid-cols-3 gap-4 w-full">
